adding SSL Encryption to all WebInterface Ports with Python OpenSSL
Its now possible to switch on SSL Encryption for each Port configured in the WebIf
As this is not a official selfsigning SSL Certificate, the Browsers aka Clients have to define an exception for this.
PLEASE NOTE: The Use of a public Private Key isnt secure. To make it secure, you have to generate your own selfsigning certifikate!
The Server Keys server.pem and the certificate cacert.pem are stored in the folder /etc/enigma2/ to have it in Backup Files